Fixed bug where scanning multiple organizations with --clean-redis resulted in cleaning redis DB each organization.
