+++ src/storage/neo4j_utils.py
@@ -1,0 +1,5 @@
+from src.config.config import Config
+
+
+def clean_graph():
+    Config.graph.clean_graph()
